end section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -51,12 +52,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
-            auction.Seller = "Test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -71,6 +73,7 @@ namespace AuctionService.Controllers
             return CreatedAtAction(nameof(GetAuctionById),new {auction.Id},newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id,UpdateAuctionDto updateAuctiondto)
         {
@@ -80,6 +83,8 @@ namespace AuctionService.Controllers
                                 .FirstOrDefaultAsync(x => x.Id == id);
             
                 if(auction == null) return NotFound();
+
+                if(auction.Seller != User.Identity.Name) return Forbid();
 
                 auction.Item.Make = updateAuctiondto.Make ?? auction.Item.Make;
                 auction.Item.Model = updateAuctiondto.Model ?? auction.Item.Model;
@@ -112,6 +117,8 @@ namespace AuctionService.Controllers
             var auction = await _context.Auctions.FindAsync(id);
 
             if(auction == null) return NotFound();
+
+            if(auction.Seller != User.Identity.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
